% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbootMISC.R, R/ggbootMV-helper.R
\name{assignGraphPar}
\alias{assignGraphPar}
\alias{assignGraphPar}
\title{Set ggbootMV graphical parameters.}
\usage{
assignGraphPar(dispPlot, checkPlot, env)

assignGraphPar(dispPlot, checkPlot, env)
}
\arguments{
\item{dispPlot, checkPlot}{logical. If \code{TRUE}, then
default values are over-ridden with appropriate values
for a display or checking plot, respectively. If both are \code{TRUE},
then dispPlot is used.}

\item{env}{environment. Environment to which to create the bindings.}

\item{displot, checkPlot}{logical. If \code{TRUE}, then
the values for certain graphical parameters given by default
or in the function call are overridden. See code of
\code{assignGraphPar} for details.}

\item{env}{environment. Environment to assign values to names in.}
}
\value{
Nothing. Assigns values to names in \code{env} directly.
}
\description{
Set the graphical parameters for ggbootMV.

\code{assignGraphPar} assigns values to the names
used for variables that determine what is plotted.
}
