% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbootMV-helper.R
\name{assignSelAxisIndVec}
\alias{assignSelAxisIndVec}
\title{Assign selected column indices to}
\usage{
assignSelAxisIndVec(selAxisLab, data, env)
}
\arguments{
\item{selAxisLab}{integer vector. Column indices of axes
to plot on biplot.}

\item{data}{numeric dataframe. Dataframe from which to take
the labels, if required.}

\item{env}{environment. Environment to which to save the results.}
}
\value{
Nothing. Assigns values to names in \code{env} directly.
}
\description{
\code{assignSelAxisIndVec} assigns the selected columns in
\code{.data} from \code{.selAxisLab} to a specified environment.
}
