% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisPred.R
\name{calcAxisPred}
\alias{calcAxisPred}
\title{Calculate axis predictivity}
\usage{
calcAxisPred(data, centre = TRUE, scale = FALSE)
}
\arguments{
\item{data}{dataframe. May}
}
\description{
Calculates the axis predictivity for a dataset.
}
