% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBootEst.R
\name{calcBCa}
\alias{calcBCa}
\title{Calculate BCa}
\usage{
calcBCa(bootObj = NULL, bootStat = NULL, origStat = NULL,
  origData = NULL, bcaA = NULL, alpha = 0.05, ci = TRUE, type = "bca",
  calcStat = NULL, B = 2000, bootPkgCI = FALSE, seed = NULL, ...)
}
\arguments{
\item{bootObj}{object of class "boot". Object generated by \code{boot::boot} function. Optional
if \code{bookPkgCI} is TRUE or if \code{bootStat} is \code{NULL}. If \code{bookPkgCI} is TRUE
and \code{bootStat} is not \code{NULL}, then \code{bootObj} is used to calculate the confidence
interval, rather than \code{bootStat}.}

\item{bootStat}{numeric vector. Bootstrap samples. Optional if \code{bookPkgCI} is TRUE. Will have no effect if
\code{bootPkgCI} is \code{TRUE} and  \code{bootObj} is not \code{NULL}.}

\item{origStat}{numeric. Original sample estimate, i.e. value of \code{calcStat}
applied to \code{origData}.}

\item{origData}{numeric vector. Original sample. Required if \code{origStat}, \code{bootStat} or
\code{bcaA} (if \code{type} == "bc" ) not given.}

\item{bcaA}{numeric. Acceleration constant.}

\item{alpha}{numeric. Dependent on \code{ci}, either 1-\code{alpha} is the coverage of the
confidence interval or \code{alpha} specifies the quantile.}

\item{ci}{logical. If \code{TRUE}, then \code{alpha} specifies the width of the confidence interval,
and an upper and lower bound are returned. If \code{FALSE}, then \code{alpha} specifies the quantile,
and a single quantile is returned.}

\item{type}{'bca' or 'bc'. Whether to return BCa or bias-corrected intervals. If type == "bc",
it sets \code{bcaA}, overriding \code{bcaA} as specified in the \code{calcBCa} arguments.}

\item{calcStat}{function. Function of a single variable to calculate for each sample.}

\item{B}{numeric. Number of bootstrap samples to generate, if required.}

\item{bootPkgCI}{logical. If TRUE, then only functions from the \code{boot} package are used
to compute the confidence interval.}

\item{seed}{numeric vector. If specified, then \code{seed(seed)} is called before
generating a bootstrap sample.}

\item{...}{arguments. Passed on to \code{boot::boot}, if run.}
}
\value{
If \code{ci} is \code{TRUE}, then a numeric vector of length two is returned with the
upper and lower bound of the 1-\code{alpha} confidence interval. If \code{ci} is \code{FALSE}, then
a single numeric is returned of the corresponding quantile.
}
\description{
Calculate BCa
}
\examples{
calcBCa( origData = c( 0.096888, 0.115579, 0.13671, 0.086497, 0.5248, 1.676752, 0.46717, 0.360975, 0.118718 ),
calcStat = mean, bootPkgCI = TRUE )
}
