% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBootEst.R
\name{calcBCaA}
\alias{calcBCaA}
\title{Calculate BCa acceleration factor}
\usage{
calcBCaA(origData, calcStat)
}
\arguments{
\item{origData}{numeric vector. Original sample.}

\item{calcStat}{function. A numeric function that takes a single numeric vector as
an argument and returns a numeric vector of length 1 as output.}
}
\value{
A numeric vector of length one that is the estimate of the acceleration factor.
}
\description{
Returns the estimated acceleration factor for the BCa confidence interval
in the one-sample non-parametric case, using the standard jack-knife.
}
\examples{
calcBCaA( c( 0.096888, 0.115579, 0.13671, 0.086497, 0.5248,
1.676752, 0.46717, 0.360975, 0.118718 ), calcStat = mean)
}
