% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBootEst.R
\name{calcBCaZ0}
\alias{calcBCaZ0}
\title{Calculate BCa bias quantile}
\usage{
calcBCaZ0(bootStat, origStat)
}
\arguments{
\item{bootStat}{numeric vector. Bootstrap sample statistics.}

\item{origStat}{numeric. Original sample statistic.}
}
\value{
A numeric vector of length one that is the estimate of the bias quantile.
}
\description{
Returns the estimated bias quantile for the BCa confidence interval.
}
\examples{
respVec = c( 0.096888, 0.115579, 0.13671, 0.086497, 0.5248,
1.676752, 0.46717, 0.360975, 0.118718 )
origStat = mean(respVec)
bootVec = boot::boot( respVec, function(x,w) mean(x[w]), R = 2e3 )$t
calcBCaZ0( bootVec, origStat )
}
