% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiplot.R
\name{calcBiplot}
\alias{calcBiplot}
\title{Bootstrap PCA biplot statistics}
\usage{
calcBiplot(dataTibble, scale, centre = TRUE, pcCompNum1 = 1,
  pcCompNum2 = 2, ellipseGroupVec = NULL, ellipseProb = 0.69,
  textAdjVal = 0.2)
}
\description{
Returns necessary statistics for PCA biplot.
}
