% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndMVBoot.R
\name{calcBootSampleGroupMVBoot}
\alias{calcBootSampleGroupMVBoot}
\title{Bootstrap location measure.}
\usage{
calcBootSampleGroupMVBoot(data, calcLoc)
}
\arguments{
\item{data}{dataframe. Numeric dataframe in wide format.}

\item{calcLoc}{function. Returns multivariate measure of location.}
}
\description{
Bootstrap location measure.
}
