% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxBiplot.R, R/calcGroupEllipse.R
\name{calcEllipse}
\alias{calcEllipse}
\alias{calcEllipse}
\title{Calculate ellipses}
\usage{
calcEllipse(data, p = 0.95)

calcEllipse(data, p = 0.95)
}
\arguments{
\item{data}{numeric dataframe}

\item{p}{numeric. Contour percentile.}
}
\value{
A matrix with points along the contour line.
}
\description{
Calculates ellipses

Calculate the \code{p}-th contour line for a distribution
using observations in \code{data}, assuming that the
distribution is bivariate normal.
}
