% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGroupEllipse.R
\name{calcGroupEllipse}
\alias{calcGroupEllipse}
\title{Calculate bivariate normal contour for groups.}
\usage{
calcGroupEllipse(data, p = 0.95)
}
\arguments{
\item{data}{dataframe. Observations in numeric columns, with grouping variable
a character column.}

\item{p}{numeric. Contour percentile.}
}
\value{
A tibble with points representing coordinates composing the contour lines
for each group, and a column denoting group for each coordinate.
}
\description{
Calculate bivariate normal contour for groups.
}
