% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndMVBoot.R, R/calcMVBoot.R
\name{calcGroupMVBoot}
\alias{calcGroupMVBoot}
\alias{calcGroupMVBoot}
\title{Bootstrap location measures for a single group.}
\usage{
calcGroupMVBoot(data, B, calcLoc)

calcGroupMVBoot(data, B, calcLoc)
}
\arguments{
\item{data}{dataframe. Numeric dataframe in wide format.}

\item{B}{integer. Number of bootstrap samples per subgroup.}

\item{calcLoc}{function. Returns multivariate measure of location.}
}
\description{
Bootstrap location measures for a single group.

Bootstrap location measures for a single group.
}
