% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMVBoot.R
\name{calcIndMVBoot}
\alias{calcIndMVBoot}
\title{Bootstrap location measure.}
\usage{
calcIndMVBoot(data, calcLoc)
}
\arguments{
\item{data}{dataframe. Numeric dataframe in wide format.}

\item{calcLoc}{function. Returns multivariate measure of location.}
}
\description{
Generate a bootstrap sample from \code{data}, and
return the
}
