% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calcMinMax}
\alias{calcMinMax}
\title{Calculate min and max}
\usage{
calcMinMax(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A named numeric vector of the minimum and maximum value.
}
\description{
Wrapper function around min and max that returns as a vector the minimum and maximum of a vector.
}
\examples{
calcMinMax(rnorm(20))
calcMinMax(1:10)
}
