% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMV.R
\name{calcMV}
\alias{calcMV}
\title{Calculate information required for plotting PCA biplot with confidence areas.}
\usage{
calcMV(data, group, B, seed, comp = 1:2, scale, calcLoc, boot = TRUE,
  save = FALSE, name = NULL, env = NULL, path = NULL, quant, trim)
}
\arguments{
\item{data}{dataframe. Only the numeric columns will be used.}

\item{group}{vector. Vector indicating group membership for each observation in \code{data}.
Will be coerced to a character vector.}

\item{B}{integer. Number of bootstrap samples per subgroup.}

\item{comp}{integer vector. Specifies the principal components
whose proportion of variation must be returned. Defaults to 1:2.}

\item{boot}{logical. If TRUE, then bootstrap is re-performed. If FALSE, then
bootstrap values are taken from object with name \code{name} in environment \code{bootList}.}

\item{save}{logical. If \code{TRUE} and \code{boot=TRUE}, then bootList is}

\item{name}{character. Name of object in \code{bootList} to add to or take to. Must not be
NULL if \code{save=TRUE}.}

\item{env}{environment. Environment to save bootstrapped values to.}

\item{path}{character. Path of to save bootstrapped values to, or load bootstrapped values from.}

\item{quant}{logical. If \code{TRUE}, then univariate 95 percent percentile bootstrap confidence
intervals are plotted.}
}
\description{
Calculate information required for plotting PCA biplot with confidence areas.
}
