% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIndMVBoot.R, R/calcMVBoot.R
\name{calcMVBoot}
\alias{calcMVBoot}
\alias{calcMVBoot}
\title{Bootstrap location measures for multiple groups.}
\usage{
calcMVBoot(data, group, B, calcLoc)

calcMVBoot(data, group, B, calcLoc)
}
\arguments{
\item{data}{dataframe. Numeric dataframe in wide format.}

\item{group}{character vector. Vector indicating group membership
for each observation in \code{data}.}

\item{B}{integer. Number of bootstrap samples per subgroup.}

\item{calcLoc}{function. Returns multivariate measure of location.}

\item{data}{dataframe. Numeric dataframe in wide format.}

\item{group}{character vector. Vector indicating group membership
for each observation in \code{data}.}

\item{calcLoc}{function. Returns multivariate measure of location.}
}
\description{
Bootstrap location measures for multiple groups.

Bootstrap location measures for multiple groups.
}
