% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBootEst.R
\name{calcP}
\alias{calcP}
\title{Calculate p-value}
\usage{
calcP(bootData, origStat, origSE = NULL, nullValue, altSide = NULL,
  method = "perc")
}
\arguments{
\item{bootData}{dataframe. Must have column named bootStat with bootstrap sample statistics.
If \code{method=='percT'}, it must also have the column bootSE, with the bootstrap estimate
standard deviation.}

\item{origStat}{numeric. Original sample estimate. Required.}

\item{origSE}{numeric. Original sample estimate standard error. Required if \code{method=='percT'}.}

\item{nullValue}{numeric. Borderline value under the null hypothesis.}

\item{altSide}{both', 'high' or 'low'. The side of the null hypothesis the alternate hypothesis
is on.}
}
\value{
The p-value as a numeric vector.
}
\description{
Calculate the p-value for a specific subgroup.
}
