% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPW.R
\name{calcPW}
\alias{calcPW}
\title{Compute pair-wise comparisons estimates and bootstrap statistics}
\usage{
calcPW(data, calcStat, trim = 0.2, B, seB, method = "percT", fdr = FALSE)
}
\arguments{
\item{data}{dataframe. Contains numeric variable named 'resp' for responses and character variable named 'group' containing factor levels.}

\item{calcStat}{function olr character. Bootstrap statistic function. Functions other than mean or median used at own risk.}

\item{B}{integer. Number of bootstrap samples per subgroup. If equal to zero, then the MBF is not calculated.}

\item{fdr}{numeric. False discovery rate.}
}
\value{
A tibble of the comparison groups and estimates of differences in \code{calcStat}, as well as the confidence intervals and MBFs if bootstrapping was performed.
}
\description{
Calculates the mean or median sample estimates of all
pair-wise differences within a factor. Can be set
compute the maximum Bayes factors (MBFs) and indicate
which of the comparisons are statistically
significant after controlling the false discovery rate, if the number of bootstrap repetition (\code{B}) is set to greater than 0.
}
