% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxBiplot.R
\name{centreAndScale}
\alias{centreAndScale}
\title{Centre and scale a dataframe.}
\usage{
centreAndScale(grid, centre, scale, message = TRUE)
}
\arguments{
\item{grid}{dataframe or matrix.}

\item{centre}{logical. If TRUE, then each column has its own
mean subtracted}

\item{scale}{logical. If TRUE, then each column is divided by
its own standard deviation. If a column has zero standard
deviation, then it is returned as is.}

\item{message}{logical. If TRUE, then}
}
\value{
A tibble or a matrix, depending on whether \code{grid}
inherited from the dataframe class or was a matrix.
}
\description{
\code{centreAndScale} allows column-wise mean centring and
standard-deviation scaling for a dataframe or matrix. The extra
features of this function are the following:
}
\details{
\itemize{
  \item If a column has zero standard deviation, then scaling
  has no effect. A message is outputted to inform the user.
  \item A dataframe may be provided with character columns. These
  are simply ignored for the purposes of centring and scaling.
  \item A dataframe or a matrix may be provided. If the
  input was a matrix, then the output is also a matrix. If the
  input inherits from the dataframe class, then the output is a
  tibble.
}
}
