% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortPW.R
\name{findSortedIndTbl}
\alias{findSortedIndTbl}
\title{Sort x- and y-columns by size of response for an individual facet.}
\usage{
findSortedIndTbl(data, plotTbl, calcStat)
}
\arguments{
\item{data}{dataframe. Has columns group, resp and facet. The column 'group' specifies individual group,
'resp' response and 'facet' overall facet.}

\item{plotTbl}{dataframe. Has columns .id, x, y and origStat. May have other columns. The column '.id' specifies overall facet,
'x' and 'y' the specific groups compared in a row, and 'origStat' how much larger the variable level in 'y' is than the variable
level in 'x' (at least for \code{ggbootPW}; for other functions it could be the other way around the code would run fine).}
}
\value{
Dataframe with columns along x- and y-axes changed appropriately.
}
\description{
This assumed that all the data is from an individual facet.
}
