% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortPW.R
\name{findXTbl}
\alias{findXTbl}
\title{Set a categorical variable as x-variable, with all remaining y-variables.}
\usage{
findXTbl(i, sortVec, plotTbl)
}
\arguments{
\item{i}{numeric. \code{sortVec[1:(length(sortVec)-i)]} are searched for as y-variables, and
\code{rev( sortVec )[i]} is set as the x-variable.}

\item{sortVec}{character vector. Order in which groups are to display along x- and y-axes. The
first group along the x-axis is the last element of \code{sortVec}.}

\item{plotTbl}{dataframe. Dataframe with columns x, y and origStat. May have other columns.
The combination \code{xVar} and \code{yVar} must only occur once in \code{plotTbl}, regardless of order.}
}
\value{
Dataframe with the columns along x and y-axes changed appropriately, for the specific variables
as set by \code{i} and \code{sortVec}.
}
\description{
Returns the rows of plotTbl with the selected group as the x-variable, and
all the y-variables that are "smaller" than this variable, as defined by \code{sortVec}.
}
\examples{
findXTbl( 1, 1:2, data.frame(x = c( 3, 1, 1 ),
y = c( 2, 2, 1 ),
origStat = c( 0.5, 0.3, -0.2 ) ) )
}
