% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbootMISC.R, R/ggbootMV-helper.R
\name{getSelAxisInd}
\alias{getSelAxisInd}
\alias{getSelAxisInd}
\title{Return the column indices with given variables names.}
\usage{
getSelAxisInd(selAxisLab, data)

getSelAxisInd(selAxisLab, data)
}
\arguments{
\item{selAxisLab}{character vector. Names of columns in \code{data} to print.}

\item{data}{dataframe.}

\item{selAxisLab}{character vector. Names of columns
whose axes must be shown on biplot. If NULL, then
all axes must be plotted, and an integer vector
from 1 to \code{ncol(data)} is returned.}

\item{data}{numeric dataframe. Dataframe from which to draw
the labels.}
}
\value{
Integer vector of columns in \code{data} to plot.
}
\description{
Return the column indices with given variables names.

\code{getSelAxisInd} returns the indices of the columns
in \code{data} whose names are in \code{selAxisLab}.
}
