% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbootMV.R
\name{ggbootMV}
\alias{ggbootMV}
\title{mv med plot info func}
\usage{
ggbootMV(data, group, B, seed = NULL, comp = 1:2, locType = "mean",
  scale = FALSE, checkPlot = FALSE, dispPlot = FALSE, labelVec = NULL,
  legendTitle = "Group", colVec = c("darkorchid1", "springgreen2",
  "maroon1", "dodgerblue", "red", "yellow3", "cyan2", "orange2"),
  addOrigPos = TRUE, pcaAxesLabSize = 2.5, origPosLabSize = 2,
  axes = TRUE, points = TRUE, pointAlpha = 1, pointSize = 1,
  ellipse = TRUE, ellAlpha = 1, ellSize = 1, quant = FALSE,
  quantAlpha = 1, quantSize = 1, density = FALSE, densAlpha = 1,
  densSize = 1, boot = TRUE, save = FALSE, name = NULL, env = NULL,
  path = NULL, textAdjVal = 0.2, axesLenRatio = NULL, qualCutoff = 0,
  selAxisLab = NULL, arrow = FALSE, arrowSize = 2, fontScaleFactor = 1,
  trim = 0.2)
}
\arguments{
\item{data}{dataframe. In wide format. Only numeric columns will be used.}

\item{group}{vector. Vector indicating group membership for each observation in \code{data}.
Will be coerced to a character vector.}

\item{B}{integer. Number of bootstrap samples per subgroup.}

\item{comp}{integer vector. Specifies the principal components
whose proportion of variation must be returned. Defaults to 1:2.}

\item{locType}{character. Specifies measure of location to use. Options are
mean, cmwed (column-wise median), oja, weisz (geometric median) and
gmed (a fast version for the geometric median).}

\item{pointSize}{numeric. Size of plot points}

\item{quant}{logical. If \code{TRUE}, then univariate 95 percent percentile bootstrap confidence
intervals are plotted.}

\item{boot}{logical. If TRUE, then bootstrap is re-performed. If FALSE, then
bootstrap values are taken from object with name \code{name} in environment \code{bootList}.}

\item{save}{logical. If \code{TRUE} and \code{boot=TRUE}, then bootList is}

\item{name}{character. Name of object in \code{bootList} to add to or take to. Must not be
NULL if \code{save=TRUE}.}

\item{env}{environment. Environment to save bootstrapped values to.}

\item{path}{character. Path of to save bootstrapped values to, or load bootstrapped values from.}

\item{textAdjVal}{numeric. Value that controls the degree to which the axis labels are shifted from
the endpoints of the axes. Higher values move the axis further right and up.}

\item{axesLenRatio}{numeric. If not NULL, then the ratio of the length of the y-axis divided by the
x-axis is made equal to \code{axesLenRatio}, where length is defined as the difference between the maximum and the
minimum value plotted along that axis. This is done by stretching the shorter axis around its mean. Useful
for making the axes of equal length.}

\item{qualCutoff}{numberic. The minimum sum of the axis predictivity across the principal components selected (by \code{comp}) required for the axis of a variable to be plotted. Only works if \code{arrow=FALSE} (this will change in future).}

\item{selAxisLab}{character vector. Names of columns in \code{data} to print.}

\item{arrow}{logical. If TRUE, then arrows instead of lines are plotted for the axes. The arrows point from
the origin to the largest observed value for its variable on the biplot axis.}

\item{arrowSize}{numeric. Size of the arrows for \code{ggplot2} to use.}

\item{fontScaleFactor}{numeric. Ratio by which to multiply the default font size.}
}
\description{
x
PCA biplot with bootstrap confidence areas.
}
\details{
Plot a PCA biplot with bootstrap confidence ares.
}
