% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbootPW.R
\name{ggbootPW}
\alias{ggbootPW}
\title{Pair-wise comparisons plot}
\usage{
ggbootPW(data, resp, group, facet = NULL, B = 100, seB = 10,
  calcStat = "mean", trim = 0.2, fdr = NULL, minEff = 0,
  groupLabVec = NULL, facetLabVec = NULL, axisLab = NULL, ncol = 2,
  facetOrderVec = NULL, rotX = FALSE, method = "percT",
  facetScale = "fixed", text = TRUE, textSize = 3, plotTblName = NULL,
  plot = TRUE, fillLabVec = NULL, fillColVec = NULL, sort = TRUE)
}
\arguments{
\item{data}{dataframe. Has one row per observation and the column names specified below.}

\item{resp}{unquoted expression. Response variable name.}

\item{group}{unquoted expression. Group variable name.}

\item{facet}{unquoted expression. Facet variable name.}

\item{B}{integer. Number of bootstrap repetitions. If greater than 0, the plot displays the MBF rather than the estimate of the difference. If 0, then no bootstrapping is performed and the sample values for \code{calcStat} are displayed.}

\item{calcStat}{function olr character. Bootstrap statistic function. Functions other than mean or median used at own risk.}

\item{fdr}{numeric. False discovery rate.}

\item{minEff}{numeric. Minimum effect size. Default is 0.}

\item{groupLabVec}{named character vector. Labels for grouping vector. Optional.}

\item{axisLab}{character. X- and y-axis titles.}

\item{ncol}{integer. Number of facet variable columns.}

\item{rotX}{logical. If TRUE, x-axis titles are rotated 90 degrees.}

\item{facetScale}{'fixed' or 'free'. Input for scale argument of \code{ggboot::facet_wrap}. Forced to \code{'free'} if \code{sort=TRUE}.}

\item{text}{logical. If TRUE, the difference estimates or MBFs are written over each grid point.}

\item{textSize}{numeric. Size of text.}

\item{plotTblName}{character. If not NULL, then the dataframe used to plot is saved to global environment with name \code{plotTblName}.}

\item{plot}{logical. If TRUE, then a plot is outputted.}

\item{scale}{logical. Whether to sort x- and y-axes text by size of response. Default is \code{TRUE}.}
}
\description{
Plot the estimates or maximum Bayes factors (MBFs) pair-wise comparisons in terms of a summary statistic between all levels of a factor. BCa bootstrapping is used to obtain the maximum Bayes factor.
}
