% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisPred.R
\name{plotCompQual}
\alias{plotCompQual}
\title{Component-wise axis predictivity plot.}
\usage{
plotCompQual(data, pc1, pc2, colVarName = "Cytokine\\nCombination",
  colVec = cytComboColVec, scale)
}
\description{
Outputs a plot of the component-wise axis predictivity.
}
