% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axisPred.R
\name{plotCumQual}
\alias{plotCumQual}
\title{Plot cumulative axis predctivity}
\usage{
plotCumQual(data, centre = TRUE, scale = FALSE, legendTitle = "Variable",
  legendLab = NULL, legendCol = NULL)
}
\arguments{
\item{data}{dataframe. Named numeric columns in wide format required.}

\item{centre}{logical. If TRUE, then each column has its own
mean subtracted}

\item{scale}{logical. If TRUE, then each column is divided by
its own standard deviation. If a column has zero standard
deviation, then it is returned as is.}

\item{legendTitle}{character. Title of legend. Default is "Variable".}

\item{legendLab}{named character vector. If NULL, then the labels used are
the names of the columns and "mean". If not NULL, then the names of the
labels must be the names of the columns in data, plus "mean".}

\item{legendCol}{character vector. Colours to be used. May be named using
the names of the columns in data, as well as "mean".}
}
\value{
Plot.
}
\description{
Plot of the cumulative axis predictivity.
}
